Add GenericStorage in sov-state
